@@ -13,7 +13,7 @@ plugins {
     id 'java'
     id 'java-library'
     id 'jacoco'
-    id 'me.champeau.gradle.japicmp' version '0.4.1'
+    id 'me.champeau.gradle.japicmp' version '0.2.9'
 }
 
 repositories {
@@ -81,7 +81,7 @@ project.afterEvaluate {
 
 ext {
     //baselineCompareVersion = '1.5.0'
-    testInJavaVersions = [8, 11, 17, 21]
+    testInJavaVersions = [11, 17, 21]
 }
 
 jacocoTestReport {
@@ -93,7 +93,7 @@ jacocoTestReport {
 
 java {
     toolchain {
-        languageVersion = JavaLanguageVersion.of(8)
+        languageVersion = JavaLanguageVersion.of(11)
     }
     // Needed because of broken gradle metadata, see https://github.com/google/guava/issues/6612#issuecomment-1614992368
     sourceSets.all {
@@ -107,8 +107,8 @@ java {
 }
 
 compileJava {
-    sourceCompatibility '1.8'
-    targetCompatibility '1.8'
+    sourceCompatibility '11'
+    targetCompatibility '11'
 }
 
 test {
@@ -120,7 +120,7 @@ test {
 }
 
 dependencies {
-    implementation 'javax.servlet:javax.servlet-api:3.1.0'
+    implementation 'jakarta.servlet:jakarta.servlet-api:5.0.0'
     implementation 'org.apache.commons:commons-lang3:3.12.0'
     implementation 'com.google.guava:guava-annotations:r03'
     implementation 'commons-codec:commons-codec:1.15'
@@ -130,11 +130,9 @@ dependencies {
     api 'com.auth0:jwks-rsa:0.22.1'
 
     testImplementation 'org.bouncycastle:bcprov-jdk15on:1.64'
-    testImplementation 'org.hamcrest:java-hamcrest:2.0.0.0'
-    testImplementation 'org.hamcrest:hamcrest-core:1.3'
-    testImplementation 'org.mockito:mockito-core:2.8.9'
-    testImplementation 'org.junit.jupiter:junit-jupiter:5.8.1'
-    testImplementation 'org.springframework:spring-test:4.3.14.RELEASE'
+    testImplementation 'org.hamcrest:hamcrest:2.2'
+    testImplementation 'org.mockito:mockito-core:5.12.0'
+    testImplementation 'org.junit.jupiter:junit-jupiter:5.10.0'
     testImplementation 'com.squareup.okhttp3:okhttp:4.11.0'
 }
 
